+++ Release Notes.md
@@ -6,14 +6,17 @@
 This release adds new classes and views:
 
 * `CornerRadiusStyle` style can be used to specify corner radii.
-* The  `cornerRadius(_ style: CornerRadiusStyle)` view modifier can be used to apply corner radius styles.
+* The  `cornerRadius(_ style:)` view modifier can be used to apply corner radius styles.
+* `FontStyle` style can be used to specify custom fonts.
+* The  `font(_ style: FontStyle)` view modifier can be used to apply custom fonts.
+* The  `font(name:, style:, weight:)` view modifier can also be used to apply custom fonts.
 * `ShadowStyle` style can be used to specify drop shadow styles.
-* The  `shadow(_ style: ShadowStyle)` view modifier can be used to apply shadow styles.
+* The  `shadow(_ style:)` view modifier can be used to apply shadow styles.
 * The `blur(...)` view modifier simplifies using the system blur styles that are specified in this library.
 
 This release also adds new extensions:
 
-* The `border(content:,width:,cornerRadius)` `View` extension can be used to apply borders with a certain content (e.g. Color.red), width and corner radius.
+* The `border(content:,width:,cornerRadius:)` `View` extension can be used to apply borders with a certain content (e.g. Color.red), width and corner radius.
 
 
 ## 0.1.0